Extract Pundit controller authorization module
