+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 - Dropped support for Ruby end-of-life versions: 2.3 (#633)
 - Dropped support for RSpec 2 (#615)
 
+### Deprecated
+
+- Deprecate `include Pundit` in favor of `include Pundit::Authorization` (#621)
+
 ## 2.1.0 (2019-08-14)
 
 ### Fixed
